Make LuaJIT compile on s390x.
I've disabled both the JIT and FFI for now. I've also stripped almost
all of the assembly out of vm_s390x.dasc, leaving only labels for the
most part. This is enough to get LuaJIT to compile but of course if
you try and run it it will explode.

The idea now is to re-add enough functionality to get a very basic
Lua program to run.